+++ abra_core/src/translate_bytecode.rs
@@ -24,7 +24,7 @@ use utils::hash::HashMap;
 use utils::hash::HashSet;
 use utils::id_set::IdSet;
 
-type OffsetTable = HashMap<NodeId, i16>;
+type OffsetTable = HashMap<NodeId, i8>;
 type MonomorphEnv = Environment<PolytypeDeclaration, Type>;
 pub(crate) type LabelMap = HashMap<Label, usize>;
 
@@ -243,7 +243,7 @@ impl Translator {
                 self.emit(st, Instr::PushNil(locals.len() as u16));
                 let mut offset_table = OffsetTable::default();
                 for (offset, node_id) in locals.iter().enumerate() {
-                    offset_table.entry(*node_id).or_insert(offset as i16);
+                    offset_table.entry(*node_id).or_insert(offset as i8);
                 }
                 for (i, item) in file.items.iter().enumerate() {
                     if let ItemKind::Stmt(stmt) = &*item.kind {
@@ -298,10 +298,10 @@ impl Translator {
                     self.emit(st, Instr::PushNil(locals_count as u16));
                     let mut offset_table = OffsetTable::default();
                     for (i, arg) in args.iter().rev().enumerate() {
-                        offset_table.entry(arg.0.id).or_insert(-(i as i16) - 1);
+                        offset_table.entry(arg.0.id).or_insert(-(i as i8) - 1);
                     }
                     for (i, local) in locals.iter().enumerate() {
-                        offset_table.entry(*local).or_insert(i as i16);
+                        offset_table.entry(*local).or_insert(i as i8);
                     }
                     let nargs = args.len();
                     st.return_stack.push(nargs as u32);
